requirements.txt
fastapi
uvicorn
openai
streamlit
pydantic
python-dotenv